develop core security 1. dao sec_* 2. jwt driven. Just now i understand benefit of jwt, to minimize query. in detail : jwt token is trusted(because encrypted with specified server key), and cointain user role (as java security approach). so controller could block or permit user access from data on JWT token

+++ coremodel/ApplicationUserToGroupAssignment.go
@@ -6,11 +6,11 @@ const tablenameUserToGroupAssignment = "sec_group_assignment"
 //ApplicationUserToGroupAssignment table: sec_group_assignment
 type ApplicationUserToGroupAssignment struct {
 	//ID id data(surrogate key), column: pk
-	ID int64 `gorm:"column:pk;AUTO_INCREMENT;primary_key" json:"id"`
+	ID int32 `gorm:"column:pk;AUTO_INCREMENT;primary_key" json:"id"`
 	//GroupID id dari group, column: group_id
-	GroupID int64 `gorm:"column:group_id" json:"groupId"`
+	GroupID int32 `gorm:"column:group_id" json:"groupId"`
 	//UserID id dari user, column: user_id
-	UserID int64 `gorm:"column:user_id" json:"userId"`
+	UserID int32 `gorm:"column:user_id" json:"userId"`
 	//User refer with column user_id
 	User ApplicationUser `gorm:"foreignkey:UserID" json:"user"`
 	//Group refer with column group_id
